config.py
from decouple import config

class Config:
    SECRET_KEY = config("SECRET_KEY")

class DevelopmentConfig(Config):
    DEBUG = True
    MYSQL_HOST = config("MYSQL_HOST")
    MYSQL_USER =  config("MYSQL_USER")
    MYSQL_PASSWORD = config("MYSQL_PASSWORD")
    MYSQL_DB = config("MYSQL_DB")


config = {
    'development': DevelopmentConfig,
    'default': DevelopmentConfig
}