requirements.txt
Flask==2.2.2
flask_login==0.6.2
flask_mysqldb==1.0.1
flask_script==2.0.6
flask_wtf==1.0.1
python-decouple==3.6
Werkzeug==2.2.2
